refactor(radio): component optimization
