+++ packages/components/radio/group.tsx
@@ -21,11 +21,11 @@ import props from './radio-group-props';
 import { RadioOptionObj, RadioOption } from './type';
 import TRadio from './radio';
 import TRadioButton from './radio-button';
-import { RadioGroupInjectionKey } from './constants';
+import { RadioGroupInjectionKey } from './consts';
 import { usePrefixClass, useCommonClassName } from '../hooks/useConfig';
 import useVModel from '../hooks/useVModel';
 import { useTNodeDefault } from '../hooks/tnode';
-import useKeyboard from './useKeyboard';
+import { useKeyboard } from './hooks';
 import { isFunction } from 'lodash-es';
 import { useMutationObserver } from '../watermark/hooks';
 import type { UseMutationObserverReturn } from '../watermark/hooks';
